feat: Qna , Notice 인증인가 기능 추가 1.유저 검증 기능 추가 2.관리자 권한 기능 추가 3.page 로직 변경 (오름차순 , 내림차순 변경가능) 4.코드 스타일 변경(줄맞춤 등)

+++ src/main/java/com/example/lineofduty/domain/qna/service/QnaService.java
@@ -2,16 +2,18 @@ package com.example.lineofduty.domain.qna.service;
 
 import com.example.lineofduty.common.exception.CustomException;
 import com.example.lineofduty.common.exception.ErrorMessage;
+import com.example.lineofduty.domain.notice.Notice;
+import com.example.lineofduty.domain.qna.QnaDto;
 import com.example.lineofduty.domain.qna.repository.QnaRepository;
-import com.example.lineofduty.domain.qna.dto.QnaDto;
 import com.example.lineofduty.domain.qna.dto.request.QnaResisterRequest;
 import com.example.lineofduty.domain.qna.dto.request.QnaUpdateRequest;
 import com.example.lineofduty.domain.qna.dto.response.QnaInquiryListResponse;
 import com.example.lineofduty.domain.qna.dto.response.QnaInquiryResponse;
 import com.example.lineofduty.domain.qna.dto.response.QnaResisterResponse;
 import com.example.lineofduty.domain.qna.dto.response.QnaUpdateResponse;
+import com.example.lineofduty.domain.user.UserDetailsImpl;
+import com.example.lineofduty.domain.qna.Qna;
 import com.example.lineofduty.domain.user.repository.UserRepository;
-import com.example.lineofduty.entity.Qna;
 import com.example.lineofduty.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -27,73 +29,85 @@ import org.springframework.transaction.annotation.Transactional;
 public class QnaService {
 
     private final QnaRepository qnaRepository;
-    private final UserRepository userRepository;
 
     // 질문 등록
-    public QnaResisterResponse qnaRegistration(Long userId,QnaResisterRequest request) {
+    public QnaResisterResponse qnaRegistration(UserDetailsImpl userDetails, QnaResisterRequest request) {
 
-        // 임시로 첫 번째 유저를 조회하여 할당 (테스트용)
-        User user = userRepository.findById(userId).orElseThrow(
-                () -> new CustomException(ErrorMessage.USER_NOT_FOUND)
-        );
+        if (userDetails.getUser().isDeleted()) {
+            throw new CustomException(ErrorMessage.USER_DELETED_NOT_FOUND);
+        }
 
-        Qna qna = new Qna(
-                request.getTitle(),
-                request.getQuestionContent(),
-                user
-        );
+        Qna qna = new Qna(request.getTitle(), request.getQuestionContent(), userDetails.getUser());
 
         qnaRepository.save(qna);
 
-        return new QnaResisterResponse(QnaDto.from(qna));
+        return QnaResisterResponse.from(qna);
     }
     //질문 단건 조회
     @Transactional(readOnly = true)
     public QnaInquiryResponse qnaInquiry(Long qnaId) {
 
-        Qna qna = qnaRepository.findById(qnaId).orElseThrow(
-                () -> new CustomException(ErrorMessage.QUESTION_NOT_FOUND)
-        );
+        Qna qna = qnaRepository.findById(qnaId)
+                .orElseThrow(() -> new CustomException(ErrorMessage.QUESTION_NOT_FOUND));
 
-        return new QnaInquiryResponse(QnaDto.from(qna));
+        return QnaInquiryResponse.from(qna);
     }
 
     //질문 목록 조회
     @Transactional(readOnly = true)
-    public QnaInquiryListResponse qnaInquiryListResponse(int page, int size, String[] sort) {
+    public QnaInquiryListResponse qnaInquiryListResponse(int page, int size, String sort) {
 
-        String sortField = sort[0];
-        String sortDirection = sort.length > 1 ? sort[1] : "desc";
-        Sort.Direction direction = Sort.Direction.fromString(sortDirection);
+        String sortProperty = "creatAt";
+        Sort.Direction sortDirection = Sort.Direction.DESC;
 
-        int pageNumber = (page > 0) ? page - 1 : 0;
+        if (sort != null && !sort.isEmpty()) {
+            String[] sortParams = sort.split(",");
+            sortProperty = sortParams[0];
+            if (sortParams.length > 1 && "asc".equalsIgnoreCase(sortParams[1])) {
+                sortDirection = Sort.Direction.ASC;
+            }
+        }
+        Pageable pageable = PageRequest.of(page, size, Sort.by(sortDirection, sortProperty));
 
-        Pageable pageable = PageRequest.of(pageNumber, size, Sort.by(direction, sortField));
+        Page<Qna> qnaPage = qnaRepository.findAll(pageable);
 
-        Page<Qna> qnaPage = qnaRepository.findAllByIsDeletedFalse(pageable);
         Page<QnaDto> qnaDtoPage = qnaPage.map(QnaDto::from);
 
         return QnaInquiryListResponse.from(qnaDtoPage);
     }
 
     //질문 수정
-    public QnaUpdateResponse qnaUpdate(Long qnaId, QnaUpdateRequest request) {
+    public QnaUpdateResponse qnaUpdate(UserDetailsImpl userDetails, Long qnaId, QnaUpdateRequest request) {
 
-        Qna qna = qnaRepository.findById(qnaId).orElseThrow(
-                ()-> new CustomException(ErrorMessage.QUESTION_NOT_FOUND)
-        );
+        if (userDetails.getUser().isDeleted()) {
+            throw new CustomException(ErrorMessage.USER_DELETED_NOT_FOUND);
+        }
+
+        Qna qna = qnaRepository.findById(qnaId)
+                .orElseThrow(()-> new CustomException(ErrorMessage.QUESTION_NOT_FOUND));
+
+        if (!userDetails.getUser().getId().equals(qna.getUser().getId())) {
+            throw new CustomException(ErrorMessage.NO_MODIFY_PERMISSION);
+        }
 
         qna.update(request.getTitle(), request.getQuestionContent());
 
-        return new QnaUpdateResponse(QnaDto.from(qna));
+        return QnaUpdateResponse.from(qna);
     }
 
     //질문 삭제
-    public void qnaDelete(Long qnaId) {
+    public void qnaDelete(UserDetailsImpl userDetails, Long qnaId) {
 
-        Qna qna = qnaRepository.findById(qnaId).orElseThrow(
-                ()-> new CustomException(ErrorMessage.QUESTION_NOT_FOUND)
-        );
+        if (userDetails.getUser().isDeleted()) {
+            throw new CustomException(ErrorMessage.USER_DELETED_NOT_FOUND);
+        }
+
+        Qna qna = qnaRepository.findById(qnaId)
+                .orElseThrow(()-> new CustomException(ErrorMessage.QUESTION_NOT_FOUND));
+
+        if (!userDetails.getUser().getId().equals(qna.getUser().getId())) {
+            throw new CustomException(ErrorMessage.NO_MODIFY_PERMISSION);
+        }
 
         qnaRepository.delete(qna);
